cosas

+++ CS2D/Assets/Scripts/Tests/MyClient.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class MyClient {
 
-    [SerializeField] private GameObject cubeClient;
+    private GameObject playerPrefab;
     private Channel channel;
     private Channel inputChannel;
     private Channel ackChannel;
@@ -16,22 +16,24 @@ public class MyClient {
     private float clientTime = 0f;
     private float packetsTime = 0f;
     private int pps;
-    private List<Actions> clientActions = new List<Actions>();
+    private List<Actions> clientActions;
     private List<ReliablePacket> packetsToSend = new List<ReliablePacket>();
     [SerializeField] int inputIndex;    
-    
-    private List<Entity> players;
-    private string myId;
-    private string id;
-    private int lastRemoved = 0;
+    private Dictionary<int, GameObject> players;
+    private int id;
+    private int lastRemoved;
 
 
-    public MyClient(GameObject cubeClient, Channel channel, Channel inputChannel, Channel ackChannel, int pps) {
-        this.cubeClient = cubeClient;
+    public MyClient(GameObject playerPrefab, Channel channel, Channel inputChannel, Channel ackChannel, int pps, int id) {
+        this.playerPrefab = playerPrefab;
         this.channel = channel;
         this.inputChannel = inputChannel;
         this.ackChannel = ackChannel;
         this.pps = pps;
+        this.players = new Dictionary<int, GameObject>();
+        this.id = id;
+        this.lastRemoved = 0;
+        this.clientActions = new List<Actions>();
     }
 
     public void UpdateClient() 
@@ -47,16 +49,9 @@ public class MyClient {
         var packet = channel.GetPacket();
 
         if (packet != null) {
-            List<CubeEntity> cubeEntities = new List<CubeEntity>();
-            foreach (Entity entity in players)
-            {
-                var cubeEntity = new CubeEntity(entity.gameObject, entity.id);   
-                cubeEntities.Add(cubeEntity);   
-            }
-            var snapshot = new Snapshot(-1, cubeEntities);
-            var buffer = packet.buffer;
-
-            snapshot.Deserialize(buffer);
+            CubeEntity cubeEntity = new CubeEntity(players[id]);
+            Snapshot snapshot = new Snapshot(cubeEntity);
+            snapshot.Deserialize(packet.buffer);
 
             int size = interpolationBuffer.Count;
             if(size == 0 || snapshot.packetNumber > interpolationBuffer[size - 1].packetNumber) {
@@ -76,10 +71,25 @@ public class MyClient {
         }
     }
 
-    public void AddClient(Entity entity) 
+    public void AddClient(int playerId, CubeEntity cubeEntity) 
     {
-        myId = entity.id;
-        players.Add(entity);
+        if (!players.ContainsKey(playerId))
+        {
+            if (id == 1)
+            {
+                if (id == playerId)
+                {
+                    Debug.Log("Spawning own with id = " + playerId);
+                    Spawn(cubeEntity);
+                }
+                else 
+                {
+                    Debug.Log("Spawning player with id = " + playerId);
+                    SpawnPlayer(playerId, cubeEntity);
+                }
+            }
+        }
+        //Send ACK
     }
     private void SendReliablePacket(Packet packet, float timeout)
     {
@@ -110,7 +120,7 @@ public class MyClient {
     {
         inputIndex += 1;
         var action = new Actions(
-            myId,
+            id,
             inputIndex, 
             Input.GetKeyDown(KeyCode.Space), 
             Input.GetKeyDown(KeyCode.LeftArrow), 
@@ -119,7 +129,6 @@ public class MyClient {
         );
 
         clientActions.Add(action);
-
         var packet = Packet.Obtain();
         
         packet.buffer.PutInt(clientActions.Count);
@@ -142,7 +151,6 @@ public class MyClient {
             }
         }
     }
-
     private void CheckIfReliablePacketReceived(int index)
     {
         List<int> toRemove = new List<int>();
@@ -164,12 +172,14 @@ public class MyClient {
     }
     private void GetServerACK() {
         var packet = ackChannel.GetPacket();
-
         // Capaz tiene que ser un while
         if (packet != null) {
             int inputIndex = packet.buffer.GetInt();
-            clientActions.RemoveRange(0, inputIndex - lastRemoved -1);
-            lastRemoved = inputIndex;
+            if(lastRemoved < inputIndex) 
+            {
+                clientActions.RemoveRange(0, inputIndex - lastRemoved -1);
+                lastRemoved = inputIndex;
+            }
             CheckIfReliablePacketReceived(inputIndex);
         }
     }
@@ -185,4 +195,20 @@ public class MyClient {
         }
     }
 
+
+    public void Spawn(CubeEntity clientCube)
+    {
+            Debug.Log("Spawning own2 clientId = " + id);
+            Vector3 position = clientCube.position;
+            Quaternion rotation = Quaternion.Euler(clientCube.eulerAngles);
+            players[id] = Object.Instantiate(playerPrefab, position, rotation) as GameObject;
+    }
+    
+    public void SpawnPlayer(int playerId, CubeEntity playerCube)
+    {
+        Vector3 position = playerCube.position;
+        Quaternion rotation = Quaternion.Euler(playerCube.eulerAngles);
+        GameObject player = GameObject.Instantiate(playerPrefab, position, rotation) as GameObject;
+        players[playerId] = player;
+    }
 }
